📝 Rename : 파일명 변경
코드리뷰 사항 반영
