+++ src/pages/MainPage/index.jsx
@@ -1,6 +1,6 @@
 import React, { useState } from 'react';
 import './MainPage.scss';
-import Header from '../../components/Header/Header';
+import Header from '../../components/Header';
 import { FaRegStar, FaStar } from 'react-icons/fa';
 import { FiSearch } from 'react-icons/fi';
 import { useNavigate } from 'react-router-dom';
